requirements.txt
cached-property==1.5.2
cycler==0.11.0
decorator==4.4.2
h5py==3.1.0
imageio==2.15.0
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.1.3
natsort==7.1.0
networkx==2.5.1
numpy==1.19.5
opencv-python==4.3.0.36
Pillow==8.3.2
protobuf==3.19.4
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.5.4
six==1.16.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
torch==1.2.0
torchvision==0.4.0
